+++ lizard_progress/models.py
@@ -7,42 +7,40 @@ from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
 
-import datetime
-import functools
-import json
-import logging
-import os
-import random
-import shutil
-import string
-
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.gis.db import models
+from django.contrib.gis.gdal import DataSource
 from django.contrib.gis.geos import LineString
 from django.contrib.gis.geos import MultiLineString
 from django.contrib.gis.geos import fromstr
-from django.contrib.gis.gdal import DataSource
+from django.contrib.gis.measure import D
 from django.contrib.sites.models import Site
 from django.core.exceptions import PermissionDenied
-from django.db.models.signals import post_save
 from django.core.urlresolvers import reverse
 from django.db import connection
+from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.http import HttpRequest
 from django.template.defaultfilters import slugify
 from django.utils.functional import cached_property
 from django.utils.translation import ugettext_lazy as _
-
 from jsonfield import JSONField
-
-import lizard_progress.specifics
 from lizard_progress.email_notifications import notify
-from lizard_progress.email_notifications.models import NotificationType
 from lizard_progress.email_notifications.models import NotificationSubscription
+from lizard_progress.email_notifications.models import NotificationType
 from lizard_progress.util import directories
 from lizard_progress.util import geo
+import datetime
+import functools
+import json
+import lizard_progress.specifics
+import logging
+import os
+import random
+import shutil
+import string
 
 RDNEW = 28992
 SRID = RDNEW
@@ -547,6 +545,8 @@ class Location(models.Model):
         (LOCATION_TYPE_DRAIN, ) * 2,
     )
 
+    CLOSE_BY_DISTANCE = 10  # m. For the multiple projects graph.
+
     # A location / scheduled measurement in an activity
     activity = models.ForeignKey('Activity', null=True)
 
@@ -573,6 +573,10 @@ class Location(models.Model):
     # type, because they're all exported to the same shapefile.
     the_geom = models.GeometryField(null=True, srid=SRID)
     is_point = models.BooleanField(default=True)
+
+    # This slight denormalisation is necessary for the Geoserver views,
+    # otherwise they become insance.
+    one_measurement_uploaded = models.BooleanField(default=False)
 
     # Any extra known information about the location
     information = JSONField(null=True, blank=True)
@@ -665,6 +669,20 @@ class Location(models.Model):
     @property
     def all_expected_attachments_present(self):
         return not self.missing_attachments().exists()
+
+    def close_by_locations_of_same_organisation(self):
+        """Return a queryset of complete locations close to this one from the
+        same organisation.
+
+        """
+        if not self.the_geom:
+            return Location.objects.empty()
+
+        return Location.objects.filter(
+            activity__project__organization=self.activity.project.organization,
+            the_geom__distance_lte=(
+                self.the_geom, D(m=self.CLOSE_BY_DISTANCE)),
+            complete=True)
 
     def set_completeness(self):
         self.complete = (
@@ -937,9 +955,7 @@ class Activity(ProjectActivityMixin, models.Model):
         if not self.measurement_type.can_be_displayed:
             return
 
-        from lizard_progress.changerequests.models import Request
-        for request in self.request_set.filter(
-                request_status=Request.REQUEST_STATUS_OPEN):
+        for request in self.request_set.all():
             yield request.map_layer()
 
         from lizard_progress.util.workspaces import MapLayer
@@ -1065,7 +1081,7 @@ class ExpectedAttachment(models.Model):
         filename = os.path.basename(path)
 
         try:
-            expected_attachment = cls.objects.get(
+            expected_attachment = cls.objects.distinct().get(
                 measurements__location__activity=activity,
                 filename__iexact=filename)
         except cls.DoesNotExist:
@@ -1267,10 +1283,6 @@ class Measurement(models.Model):
         - Deletes ExpectedAttachments that aren't connected to any
           Measurements anymore due to the previous item.
 
-        Return a boolean that is True if all expected attachments for
-        this measurement have been uploaded, or False if there are
-        still expected attachments that weren't.
-
         """
         filenames = set(filenames)
 
@@ -1291,8 +1303,9 @@ class Measurement(models.Model):
             try:
                 # If the attachment already exists and is attached to this,
                 # then that is OK.
-                expected_attachment = ExpectedAttachment.objects.get(
+                ExpectedAttachment.objects.get(
                     filename=filename, measurements=self)
+                continue
             except ExpectedAttachment.DoesNotExist:
                 try:
                     # If an attachment with the same filename is
@@ -1310,9 +1323,10 @@ class Measurement(models.Model):
                     # this is probably unintended by the user anyway,
                     # he just gave two files on two different dates the
                     # same name). So then we raise an exception.
-                    expected_attachment = ExpectedAttachment.objects.get(
-                        filename=filename,
-                        measurements__location__activity=activity)
+                    expected_attachment = (
+                        ExpectedAttachment.objects.distinct().get(
+                            filename=filename,
+                            measurements__location__activity=activity))
                     if expected_attachment.uploaded:
                         raise AlreadyUploadedError(filename)
 
@@ -1527,8 +1541,8 @@ class UploadedFile(models.Model):
         """This will be turned into JSON to send to the UI."""
         return {
             'id': self.id,
-            'project_id': self.activity.project.id,
-            'activity_id': self.activity.id,
+            'project_id': self.activity.project_id,
+            'activity_id': self.activity_id,
             'uploaded_by': self.get_uploaded_by_name(),
             'uploaded_at': self.uploaded_at.strftime("%d/%m/%y %H:%M"),
             'filename': os.path.basename(self.path),
@@ -1728,9 +1742,11 @@ class ExportRun(models.Model):
         directory = directories.exports_dir(self.activity)
         return os.path.join(
             directory,
-            "{project}-{activityid}-{exporttype}.{extension}").format(
+            "{project}-{activityid}-{activity}-{exporttype}.{extension}"
+        ).format(
             project=self.activity.project.slug,
             activityid=self.activity.id,
+            activity=directories.clean(self.activity.name),
             exporttype=self.exporttype,
             extension=extension).encode('utf8')